Lib finally working and almost done!
Added functions like sort by max to min the scores, find a score typing the name, add new scores to the table, and more. At the first i use structs to make the table, but that bring me errors. To fix it i use simple arrays. (With the lib there is a project to try his features)